refactor: create switch base and update widgets
